Add WKB Reader

+++ lib/src/Bytes.cpp
@@ -14,7 +14,7 @@ namespace geopackage {
         bytes.insert(std::end(bytes), std::begin(value), std::end(value));
     }
     
-    void Bytes::putInt(int value) {
+    void Bytes::putInt(int32_t value) {
          if(endian == Endian::BIG) {
             bytes.push_back(static_cast<std::byte>(value >> 24));
             bytes.push_back(static_cast<std::byte>(value >> 16));
@@ -56,20 +56,22 @@ namespace geopackage {
         putLong(longValue);
     }
 
-    int Bytes::getInt() {
+    int32_t Bytes::getInt() {
         int value;
         if(endian == Endian::BIG) {
-            value = static_cast<int>(
-            ((bytes[position + 0] & std::byte{0xff}) << 24) |
-            ((bytes[position + 1] & std::byte{0xff}) << 16) |
-            ((bytes[position + 2] & std::byte{0xff}) << 8)  |
-            ((bytes[position + 3] & std::byte{0xff})));
+            value = static_cast<int32_t>(
+                (static_cast<int32_t>((bytes[position + 0] & std::byte{0xff})) << 24) |
+                (static_cast<int32_t>((bytes[position + 1] & std::byte{0xff})) << 16) |
+                (static_cast<int32_t>((bytes[position + 2] & std::byte{0xff})) << 8)  |
+                (static_cast<int32_t>((bytes[position + 3] & std::byte{0xff})))
+            );
         } else {
-            value = static_cast<int>(
-            ((bytes[position + 3] & std::byte{0xff}) << 24) |
-            ((bytes[position + 2] & std::byte{0xff}) << 16) |
-            ((bytes[position + 1] & std::byte{0xff}) << 8)  |
-            ((bytes[position + 0] & std::byte{0xff})));    
+            value = static_cast<int32_t>(
+                (static_cast<int32_t>((bytes[position + 3] & std::byte{0xff})) << 24) |
+                (static_cast<int32_t>((bytes[position + 2] & std::byte{0xff})) << 16) |
+                (static_cast<int32_t>((bytes[position + 1] & std::byte{0xff})) << 8)  |
+                (static_cast<int32_t>((bytes[position + 0] & std::byte{0xff})))
+            );
         }
         position = position + 4;
         return value;
@@ -78,23 +80,23 @@ namespace geopackage {
     int64_t Bytes::getLong() {
         int64_t value;
         if(endian == Endian::BIG) {
-            value = static_cast<int64_t>(bytes[position + 0]) << 56
-                | static_cast<int64_t>(bytes[position + 1] & std::byte{0xff}) << 48
-                | static_cast<int64_t>(bytes[position + 2] & std::byte{0xff}) << 40
-                | static_cast<int64_t>(bytes[position + 3] & std::byte{0xff}) << 32
-                | static_cast<int64_t>(bytes[position + 4] & std::byte{0xff}) << 24
-                | static_cast<int64_t>(bytes[position + 5] & std::byte{0xff}) << 16
-                | static_cast<int64_t>(bytes[position + 6] & std::byte{0xff}) <<  8
-                | static_cast<int64_t>(bytes[position + 7] & std::byte{0xff});
+            value = static_cast<int64_t>(bytes[position + 0]) << 56 |
+                static_cast<int64_t>(bytes[position + 1] & std::byte{0xff}) << 48 |
+                static_cast<int64_t>(bytes[position + 2] & std::byte{0xff}) << 40 |
+                static_cast<int64_t>(bytes[position + 3] & std::byte{0xff}) << 32 |
+                static_cast<int64_t>(bytes[position + 4] & std::byte{0xff}) << 24 |
+                static_cast<int64_t>(bytes[position + 5] & std::byte{0xff}) << 16 |
+                static_cast<int64_t>(bytes[position + 6] & std::byte{0xff}) <<  8 |
+                static_cast<int64_t>(bytes[position + 7] & std::byte{0xff});
         } else {
-            value = static_cast<int64_t>(bytes[position + 7]) << 56
-                | static_cast<int64_t>(bytes[position + 6] & std::byte{0xff}) << 48
-                | static_cast<int64_t>(bytes[position + 5] & std::byte{0xff}) << 40
-                | static_cast<int64_t>(bytes[position + 4] & std::byte{0xff}) << 32
-                | static_cast<int64_t>(bytes[position + 3] & std::byte{0xff}) << 24
-                | static_cast<int64_t>(bytes[position + 2] & std::byte{0xff}) << 16
-                | static_cast<int64_t>(bytes[position + 1] & std::byte{0xff}) <<  8
-                | static_cast<int64_t>(bytes[position + 0] & std::byte{0xff});
+            value = static_cast<int64_t>(bytes[position + 7]) << 56 |
+                static_cast<int64_t>(bytes[position + 6] & std::byte{0xff}) << 48 |
+                static_cast<int64_t>(bytes[position + 5] & std::byte{0xff}) << 40 |
+                static_cast<int64_t>(bytes[position + 4] & std::byte{0xff}) << 32 |
+                static_cast<int64_t>(bytes[position + 3] & std::byte{0xff}) << 24 |
+                static_cast<int64_t>(bytes[position + 2] & std::byte{0xff}) << 16 |
+                static_cast<int64_t>(bytes[position + 1] & std::byte{0xff}) <<  8 |
+                static_cast<int64_t>(bytes[position + 0] & std::byte{0xff});
         }
         position = position + 8;
         return value;
